Version 1.2.1 Refactor: fix comments

+++ ewm-service/src/main/java/ru/practicum/event/dto/EventShortDto.java
@@ -20,7 +20,7 @@ public class EventShortDto {
     private String annotation;
     private CategoryDto category;
     private UserShortDto initiator;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     private Boolean paid;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = DATE_TIME_PATTERN)
