+++ ewm-service/src/main/java/ru/practicum/request/dto/ParticipationRequestDto.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.request.enums.Status;
+import ru.practicum.request.dto.EventRequestStatusUpdateDto.Status;
 
 import java.time.LocalDateTime;
 
@@ -17,7 +17,7 @@ public class ParticipationRequestDto {
     private Long id;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = DATE_TIME_PATTERN)
     private LocalDateTime created;
-    private Integer event;
-    private Integer requester;
+    private Long event;
+    private Long requester;
     private Status status;
 }
